pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "kartuli-bot"
version = "0.1.0"
description = "Telegram bot for Georgian vocabulary with Leitner repetition"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "aiogram>=3.15.0",
  "python-dotenv>=1.0.1",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"
